+++ src/barcoder/barcode-write.cc
@@ -33,11 +33,11 @@ int main( int argc, char *argv[] )
 
   /* open file and check for sane length */
   File input { argv[ 1 ] };
-  const uint32_t width = paranoid_atoi( argv[ 2 ] );
-  const uint32_t height = paranoid_atoi( argv[ 3 ] );
-  const uint64_t frame_length = width * height * sizeof( RGBPixel );
+  const uint16_t width = paranoid_atoi( argv[ 2 ] );
+  const uint16_t height = paranoid_atoi( argv[ 3 ] );
+  const size_t frame_length = width * height * sizeof( RGBPixel );
 
-  const unsigned int frame_count = input.size() / (uint64_t)frame_length;
+  const size_t frame_count = input.size() / (uint64_t)frame_length;
   if ( input.size() != frame_count * frame_length ) {
     throw runtime_error( "file size is not multiple of frame size" );
   } else {
